処理をゲッターに移動した。イベント詳細でもラベルが付くようにした
