+++ yeahcheese/resources/js/components/BaseEvent.vue
@@ -5,8 +5,8 @@
         {{ eventInfo.name }}
         <span
           class="badge"
-          :class="getLabelByDeadline(strDel(eventInfo.start_date), strDel(eventInfo.end_date)).class"
-        >{{ getLabelByDeadline(strDel(eventInfo.start_date), strDel(eventInfo.end_date)).text }}</span>
+          :class="this.getLabelByDeadline(eventInfo.start_date, eventInfo.end_date).class"
+        >{{ this.getLabelByDeadline(eventInfo.start_date, eventInfo.end_date).text }}</span>
       </h3>
       <router-link
         class="btn btn-outline-primary float-right"
@@ -35,6 +35,7 @@
 </template>
 
 <script>
+import { mapGetters } from "vuex";
 export default {
   name: "BaseEvent",
   props: {
@@ -44,46 +45,13 @@ export default {
     }
   },
   computed: {
+    ...mapGetters({
+      getLabelByDeadline: "events/getLabelByDeadline"
+    }),
     alt() {
       return function(id) {
         return this.eventInfo.name + "の写真" + id;
       };
-    },
-    // イベントの公開開始日と終了日を取得し、公開開始前、開始後、終了後でラベルを変更させる
-    getLabelByDeadline() {
-      return function(start_date, end_date) {
-        let today = this.getToday();
-        // 公開開始前
-        if (today < parseInt(start_date)) {
-          return { class: "badge-secondary", text: "公開前" };
-        }
-        // 公開終了後
-        if (parseInt(end_date) < today) {
-          return { class: "badge-danger", text: "公開終了" };
-        }
-        // 公開中
-        return { class: "badge-success", text: "公開中" };
-      };
-    }
-  },
-  methods: {
-    getToday() {
-      let now = new Date();
-      let y = now.getFullYear();
-      let m = now.getMonth() + 1;
-      let d = now.getDate();
-      if (m < 10) {
-        m = "0" + m;
-      }
-      if (d < 10) {
-        d = "0" + d;
-      }
-      return y + m + d;
-    },
-    // イベントのstart_dateとend_dateのハイフンを消す
-    strDel(str) {
-      let res = str.slice(0, 4) + str.slice(5, 7) + str.slice(8);
-      return res;
     }
   }
 };
